Refactor ClusterServiceImpl into multiple classes
Refactored ClusterServiceImpl into multiple classes, grouping by functionality.
Like; join fucntions as a group, heartbeat & master confirmation as a group etc.
